requirements.txt
neo4j
neo4j-graphrag[openai]
openai
python-dotenv
graphrag
numpy
